inference_worker/requirements.txt
accelerate==0.25.0
diffusers==0.24.0
huggingface_hub==0.20.2
transformers==4.36.2
peft==0.7.1

pika==1.3.2
wandb==0.22.2
python-dotenv==1.1.1
google-cloud==0.34.0
google-cloud-storage==3.4.1
numpy==1.26.3
python-json-logger==4.0.0

opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-pika==0.59b0
opentelemetry-instrumentation-requests==0.59b0
deprecated==1.3.1

protobuf==3.20.3
pydantic-settings==2.11.0

celery==5.5.3
sqlalchemy==2.0.44
amqp==5.3.1
opentelemetry-instrumentation-sqlalchemy==0.59b0
psycopg2-binary==2.9.11

black==25.9.0
isort==7.0.0